requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
attrs==25.1.0
audioread==3.0.1
Bottleneck==1.4.2
contourpy==1.3.2
cycler==0.12.1
datasets==3.2.0
descript-audio-codec
decorator==5.2.1
decord==0.6.0
dill==0.3.8
et_xmlfile==2.0.0
filelock==3.17.0
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2024.9.0
grpcio==1.70.0
h5py==3.13.0
huggingface-hub==0.31.2
imageio==2.37.0
intel-cmplr-lib-ur==2025.1.1
intel-openmp==2025.1.1
Jinja2==3.1.5
joblib==1.4.2
jsonpointer==2.1
kiwisolver==1.4.8
lap==0.5.12
lazy_loader==0.4
librosa==0.11.0
lightning-utilities==0.14.3
llvmlite==0.44.0
loguru==0.7.3
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.1
mkl==2025.1.0
mkl-service==2.4.2
mkl_fft==1.3.14
mkl_random==1.2.10
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
ninja==1.11.1.4
numba==0.61.2
numexpr==2.10.2
numpy==1.26.4
opencv-contrib-python==4.7.0.68
opencv-python==4.7.0.68
openpyxl==3.1.5
pandas==2.2.3
pillow==11.2.1
pooch==1.8.2
propcache==0.2.1
pyarrow==19.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.1.post0
pytz==2024.2
PyYAML==6.0.2
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
setuptools==69.5.1
six==1.17.0
soundfile==0.13.1
soxr==0.5.0.post1
sympy==1.14.0
tabulate==0.9.0
tbb==2022.1.0
tcmlib==1.3.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tifffile==2025.3.30
torch==2.7.0
torch-levenberg-marquardt==1.2.2
torchaudio
torchmetrics
torchvision
tqdm 
typing_extensions==4.12.2
tzdata==2025.1
umf==0.10.0
Werkzeug==3.1.3
wheel==0.43.0
win32_setctime==1.2.0
xmltodict==0.14.2
xxhash==3.5.0
yarl==1.18.3
